chore: reflect 데이터 호출 방식 변경 및 reflect.ts 파일로 이동
